requirements.txt
flask
redis
numpy
pandas
mysql-connector-python